Update: move HtmlCommentBlockFilter to comment block filter module
